wip. Moved everything QTDesktopVideoDevice to DesktopVideoDevice (not yet functional)
