+++ src/lib/app/RvCommon/DesktopVideoDevice.cpp
@@ -8,6 +8,9 @@
 #ifdef PLATFORM_WINDOWS
 #include <GL/glew.h>
 #include <GL/wglew.h>
+#include <Wingdi.h>
+#include <Shlwapi.h>
+#include <lcms2.h>
 #endif
 
 #include <RvCommon/DesktopVideoDevice.h>
@@ -31,28 +34,71 @@ namespace Rv
     using namespace TwkApp;
 
     DesktopVideoDevice::DesktopVideoDevice(VideoModule* m,
-                                           const std::string& name,
-                                           const QTGLVideoDevice* share)
+                                           const std::string& name, int screen,
+                                           const QTGLVideoDevice* glViewShared)
         : GLBindableVideoDevice(m, name, ImageOutput | ProvidesSync | SubWindow)
         , m_viewDevice(0)
-        , m_share(share)
+        , m_share(glViewShared)
         , m_stereoMode(Mono)
         , m_vsync(true)
         , m_videoFormatIndex(0)
         , m_dataFormatIndex(0)
         , m_syncing(false)
+        , m_screen(screen)
+        , m_translator(0)
+        , m_view(0)
     {
         m_glGlobalState = new GLState();
         m_glGlobalState->useGLProgram(textureRectGLProgram());
+
+        const QList<QScreen*> screens = QGuiApplication::screens();
+        if (screen < 0 || screen >= screens.size())
+        {
+            // Handle invalid screen index.
+            // TODO_QT
+            std::cout << "Screen index is out of range" << std::endl;
+        }
+
+        const QRect rect = screenGeometry();
+
+        ostringstream str;
+        str << rect.width() << " x " << rect.height();
+        m_videoFormats.push_back(
+            VideoFormat(rect.width(), rect.height(), 1.0, 1.0, 0.0, str.str()));
+        addDefaultDataFormats();
     }
 
-    DesktopVideoDevice::~DesktopVideoDevice() { delete m_glGlobalState; }
+    DesktopVideoDevice::~DesktopVideoDevice()
+    {
+        close();
+        delete m_glGlobalState;
+    }
 
-    void DesktopVideoDevice::redraw() const { ScopedLock lock(m_mutex); }
+    void DesktopVideoDevice::redraw() const
+    {
+        if (m_view)
+        {
+            ScopedLock lock(m_mutex);
+            QSize s = m_view->size();
+            m_view->update();
+        }
+    }
 
     void DesktopVideoDevice::redrawImmediately() const
     {
-        ScopedLock lock(m_mutex);
+        if (m_view && m_view->isVisible())
+        {
+            ScopedLock lock(m_mutex);
+            TWK_GLDEBUG;
+            m_view->update();
+            TWK_GLDEBUG;
+        }
+        else
+        {
+            TWK_GLDEBUG;
+            redraw();
+            TWK_GLDEBUG;
+        }
     }
 
     void DesktopVideoDevice::blockUntilSyncComplete() const
@@ -111,6 +157,73 @@ namespace Rv
         // Copy the FBO (first time = black) into the FBO of the QOpenGLWidget.
         local_fbo->copyTo(defaultFBO);
         local_fbo->unbind();
+    }
+
+    void DesktopVideoDevice::open(const StringVector& args)
+    {
+        //
+        //  always make the fullscreen device synced
+        //
+
+        // NOTE_QT: QSurfaceFormat replace QGLFormat.
+        TWK_GLDEBUG;
+
+        QSurfaceFormat fmt = shareDevice()->widget()->format();
+        fmt.setSwapInterval(m_vsync ? 1 : 0);
+        ScreenView* s =
+            new ScreenView(fmt, 0, shareDevice()->widget(), Qt::Window);
+        setWidget(s);
+
+        TWK_GLDEBUG;
+
+        auto viewDevice = new QTGLVideoDevice(0, "local view", s);
+        setViewDevice(viewDevice);
+
+        TWK_GLDEBUG;
+
+        const QList<QScreen*> screens = QGuiApplication::screens();
+        QRect g = screenGeometry();
+        widget()->move(g.x(), g.y());
+        widget()->setGeometry(g);
+
+        if (useFullScreen())
+            widget()->setWindowState(Qt::WindowFullScreen);
+        else
+            widget()->setWindowState(Qt::WindowNoState);
+
+        TWK_GLDEBUG;
+
+        widget()->show();
+
+        TWK_GLDEBUG;
+
+#ifdef PLATFORM_WINDOWS
+        // Work around for QGLWidget::showFullScreen() windows specific issue
+        // introduced in Qt 5.15.3
+        widget()->setGeometry(g);
+#endif
+    }
+
+    void DesktopVideoDevice::close()
+    {
+        delete m_view;
+        delete m_viewDevice;
+        delete m_translator;
+        m_view = 0;
+        m_viewDevice = 0;
+        m_translator = 0;
+    }
+
+    void DesktopVideoDevice::setWidget(QOpenGLWidget* widget)
+    {
+        m_view = widget;
+        m_translator = new QTTranslator(this, m_view);
+    }
+
+    void DesktopVideoDevice::makeCurrent() const
+    {
+        if (m_view)
+            m_view->makeCurrent();
     }
 
     void DesktopVideoDevice::setupModelviewAndProjection(
@@ -756,6 +869,218 @@ namespace Rv
     void DesktopVideoDevice::sortVideoFormatsByWidth()
     {
         sort(m_videoFormats.begin(), m_videoFormats.end(), widthSort);
+    }
+
+    DesktopVideoDevice::ScreenView::ScreenView(const QSurfaceFormat& fmt,
+                                               QWidget* parent,
+                                               QOpenGLWidget* glViewShare,
+                                               Qt::WindowFlags flags)
+        : QOpenGLWidget(parent, flags)
+    {
+        m_glViewShare = glViewShare;
+
+        setFormat(fmt);
+    }
+
+    void DesktopVideoDevice::ScreenView::initializeGL()
+    {
+        TRACE_SCOPE("ScreenView::initializeGL")
+
+        QOpenGLWidget::initializeGL();
+
+        if (m_glViewShare && context() && context()->isValid())
+        {
+            context()->setShareContext(m_glViewShare->context());
+        }
+    }
+
+    void DesktopVideoDevice::ScreenView::paintGL()
+    {
+        TRACE_SCOPE("ScreenView::paintGL")
+
+        TWK_GLDEBUG_CTXNAME("ScreenView");
+
+        TWK_GLDEBUG;
+
+        std::cerr
+            << "SCREENVIEW FBO "
+            << QOpenGLContext::currentContext()->defaultFramebufferObject()
+            << "WFBO:" << defaultFramebufferObject() << std::endl;
+        //            glClearColor(0, 1, 0, 1);
+        //            glClear(GL_COLOR_BUFFER_BIT);
+        TWK_GLDEBUG;
+    }
+
+    //----------------------------------------------------------------------
+
+    bool DesktopVideoDevice::isOpen() const { return m_view != 0; }
+
+    bool DesktopVideoDevice::useFullScreen() const
+    {
+#ifdef DEBUG_NO_FULLSCREEN
+        return false;
+#else
+        return true;
+#endif
+    }
+
+    QRect DesktopVideoDevice::screenGeometry() const
+    {
+        const QList<QScreen*> screens = QGuiApplication::screens();
+        QRect g = screens[m_screen]->geometry();
+
+#ifdef DEBUG_NO_FULLSCREEN
+        // in the no-fullscreen debug mode, use 800x600 resolution in a simple
+        // window.
+        int offset = 30;
+        g = QRect(g.x() + offset, g.y() + offset, 800, 600);
+#endif
+        return g;
+    }
+
+    VideoDevice::Resolution DesktopVideoDevice::resolution() const
+    {
+        QRect g = screenGeometry();
+
+        return Resolution(g.width(), g.height(), 1.0, 1.0);
+    }
+
+    size_t DesktopVideoDevice::width() const
+    {
+        QRect g = screenGeometry();
+        return g.width();
+    }
+
+    size_t DesktopVideoDevice::height() const
+    {
+        QRect g = screenGeometry();
+        return g.height();
+    }
+
+    VideoDevice::VideoFormat DesktopVideoDevice::format() const
+    {
+        QRect g = screenGeometry();
+        ostringstream str;
+        str << g.width() << " x " << g.height();
+        const float pa = 1.0;
+        const float ps = 1.0;
+        const float rate = 60.0;
+        return VideoFormat(g.width(), g.height(), pa, ps, rate, str.str());
+    }
+
+    VideoDevice::Offset DesktopVideoDevice::offset() const
+    {
+        return Offset(0, 0);
+    }
+
+    VideoDevice::Timing DesktopVideoDevice::timing() const { return format(); }
+
+#ifdef PLATFORM_WINDOWS
+    TwkApp::VideoDevice::ColorProfile DesktopVideoDevice::colorProfile() const
+    {
+        //
+        // XXX The following steps are not Unicode safe
+        //
+
+        // Get the context for this screen
+        const QList<QScreen*> screens = QGuiApplication::screens();
+
+        // Ensure the screen index is valid.
+        if (m_screen < 0 || m_screen >= screens.size())
+        {
+            m_colorProfile = ColorProfile();
+            return m_colorProfile;
+        }
+
+        QScreen* targetScreen = screens[m_screen];
+        QWindow* windowOnTargetScreen = nullptr;
+        const QList<QWindow*> windows = QGuiApplication::topLevelWindows();
+
+        // Check all windows to find one on the target screen.
+        for (QWindow* window : windows)
+        {
+            if (window->screen() == targetScreen)
+            {
+                windowOnTargetScreen = window;
+            }
+        }
+
+        if (!windowOnTargetScreen)
+        {
+            // Return empty profile if no window is found on the screen.
+            m_colorProfile = ColorProfile();
+            return m_colorProfile;
+        }
+
+        HWND hwnd = reinterpret_cast<HWND>(windowOnTargetScreen->winId());
+        HDC hdc = GetDC(hwnd);
+
+        if (hdc)
+        {
+            // Look for the profile path
+            unsigned long pathLen;
+            GetICMProfile(hdc, &pathLen, NULL);
+            char* path = new char[pathLen];
+
+            if (GetICMProfile(hdc, &pathLen, path))
+            {
+                // If we found a profile lets set the type,
+                // url, and description
+
+                m_colorProfile.type = ICCProfile;
+
+                unsigned long maxLen = 2084;
+                char* url = new char[maxLen];
+                UrlCreateFromPath(path, url, &maxLen, NULL);
+                m_colorProfile.url = url;
+
+                char desc[256];
+                cmsHPROFILE profile = cmsOpenProfileFromFile(path, "r");
+                cmsGetProfileInfoASCII(profile, cmsInfoDescription, "en", "US",
+                                       desc, 256);
+                m_colorProfile.description = desc;
+
+                delete url;
+            }
+
+            delete path;
+            ReleaseDC(hwnd, hdc);
+        }
+        else
+        {
+            m_colorProfile = ColorProfile();
+        }
+
+        return m_colorProfile;
+    }
+#endif
+
+    std::vector<VideoDevice*> DesktopVideoDevice::createDesktopVideoDevices(
+        TwkApp::VideoModule* module, const QTGLVideoDevice* shareDevice)
+    {
+        std::vector<VideoDevice*> devices;
+
+        const auto screens = QGuiApplication::screens();
+        for (int screen = 0; screen < screens.size(); screen++)
+        {
+            const QScreen* w = screens[screen];
+            QString name = QString("%1 %2 %3")
+                               .arg(w->manufacturer())
+                               .arg(w->model())
+                               .arg(w->name());
+
+            DesktopVideoDevice* sd = new DesktopVideoDevice(
+                module, name.toUtf8().constData(), screen, shareDevice);
+
+            devices.push_back(sd);
+        }
+
+        return devices;
+    }
+
+    void DesktopVideoDevice::syncBuffers() const
+    {
+        redrawImmediately(); //?
     }
 
 } // namespace Rv
